Made news functionality user specific

+++ annoyed-alligators/socl_media/apps/terminal/commands.py
@@ -296,7 +296,7 @@ def news_func(request, topic: str, start_date: str = None, end_date: str = None,
 
     if page_num == 0 and article_num == 0:
         try:
-            NewsHistory.objects.latest('search_time').delete()
+            NewsHistory.objects.filter(user_id=request.user.id).latest('search_time').delete()
         except Exception:
             pass
 
@@ -305,12 +305,12 @@ def news_func(request, topic: str, start_date: str = None, end_date: str = None,
         googlenews.getpage(1)
         articles = googlenews.result()
         articles = [article for article in articles if len(article['title']) > 10]
-        db_entry = NewsHistory(user_id=1, search_topic=topic, last_fetched_count=0, news_articles=str(articles))
+        db_entry = NewsHistory(user_id=request.user.id, search_topic=topic, last_fetched_count=0, news_articles=str(articles))
         articles = articles[0:3]
         db_entry.save()
 
     else:
-        news_list = NewsHistory.objects.latest('search_time')
+        news_list = NewsHistory.objects.filter(user_id=request.user.id).latest('search_time')
         news_items = ast.literal_eval(news_list.news_articles)
         if page_num != 0:
             article_start_num = page_num * 3
